backend/pyproject.toml
[tool.poetry]
name = "skiresorts"
version = "0.1.0"
description = "Ski resort finder backend"
authors = []
packages = [{include = "skiresorts"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
strawberry-graphql = {version = "^0.262.0", extras = ["fastapi"]}
uvicorn = {version = "^0.34.0", extras = ["standard"]}
httpx = "^0.28.0"
pydantic-settings = "^2.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24.0"
respx = "^0.22.0"
ruff = "^0.15.8"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
